% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{create_website_index}
\alias{create_website_index}
\title{create_website_index}
\usage{
create_website_index(drat_repo, win_version, output_dir)
}
\arguments{
\item{drat_repo}{a}

\item{win_version}{a}

\item{output_dir}{a}
}
\description{
create_website_index
}
