% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{create_website_packages}
\alias{create_website_packages}
\title{create_website_packages}
\usage{
create_website_packages(drat_repo, win_version,
  output_dir = file.path(drat_repo, "packages"), pkgdown_base_url)
}
\arguments{
\item{drat_repo}{a}

\item{win_version}{a}

\item{output_dir}{a}

\item{pkgdown_base_url}{a}
}
\description{
create_website_packages
}
