% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{get_package_table}
\alias{get_package_table}
\title{get_package_table}
\usage{
get_package_table(drat_repo, win_version)
}
\arguments{
\item{drat_repo}{a}

\item{win_version}{a}
}
\description{
get_package_table
}
